Backend/requirements.txt
fastapi
uvicorn
bcrypt
PyJWT
email-validator
python-multipart
httpx
groq
pydantic
python-dotenv
sentence-transformers
torch
pdfminer.six
scikit-learn
